apps/guardian/pyproject.toml
[tool.poetry]
name = "offsec-guardian"
version = "0.1.0"
description = "OffSec Guardian - AI-powered threat detection agent"
authors = ["OffSec Shield Contributors"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^2.0"
httpx = "^0.25.0"
pyyaml = "^6.0"
python-dotenv = "^1.0"
toml = "^0.10"
PyJWT = {version = "^2.8.0", extras = ["crypto"]}
cryptography = "^41.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4"
pytest-cov = "^4.1"
black = "^23.0"
isort = "^5.12"
pylint = "^3.0"

[tool.poetry.scripts]
guardian = "guardian.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
